requirements.txt
gitpython
langchain
langgraph
gpt4all
langchain-huggingface
sentence-transformers
torch
faiss-cpu
langchain-text-splitters
langchain_community